Refactor errors logic to use validateProject
