+++ src/logic/__tests__/startAccountMigration.test.js
@@ -1,5 +1,3 @@
-import reduce from 'lodash-es/reduce';
-
 import {
   accountMigrationComplete,
   accountMigrationError,
@@ -10,11 +8,10 @@ import {
   userAuthenticated,
 } from '../../actions/user';
 import {migrateAccount} from '../../clients/firebase';
-import rootReducer from '../../reducers';
 import {bugsnagClient} from '../../util/bugsnag';
 import startAccountMigration from '../startAccountMigration';
 
-import {makeTestLogic} from './helpers';
+import {makeTestLogic, applyActions} from './helpers';
 
 import {
   credentialFactory,
@@ -111,11 +108,3 @@ describe('startAccountMigration', () => {
     expect(migrateAccount).not.toHaveBeenCalledWith(mockCredential);
   });
 });
-
-function applyActions(...actions) {
-  return reduce(
-    actions,
-    (state, action) => rootReducer(state, action),
-    undefined,
-  );
-}
