--- src/logic/helpers/validateSource.js
@@ -1,27 +1,0 @@
-import {validatedSource} from '../../actions/errors';
-import {getCurrentProject} from '../../selectors';
-import retryingFailedImports from '../../util/retryingFailedImports';
-
-function importValidations() {
-  return retryingFailedImports(() =>
-    import(
-      /* webpackChunkName: "mainAsync" */
-      '../../validations'
-    ),
-  );
-}
-
-export default async function validateSource(
-  {language, source, projectAttributes},
-  dispatch,
-  getState,
-) {
-  const validations = await importValidations();
-  const validate = validations[language];
-  const validationErrors = await validate(source, projectAttributes);
-  const state = getState();
-  const currentProject = getCurrentProject(state);
-  if (currentProject.sources[language] === source) {
-    dispatch(validatedSource(language, validationErrors));
-  }
-}
